🚀 feat: update dependencies (#2)
* 🚀 feat: update dependencies

- Added autoprefixer version 10.4.18
- Added daisyui version 4.7.2
- Added postcss version 8.4.35
- Added tailwindcss version 3.4.1
- Updated eslint version to 8.56.0
- Updated typescript version to 5.2.2
- Updated vite version to 5.1.4

* 🎨 style: update login form UI and remove unused code

The changes in this commit update the UI of the login form in the client application. The changes include adding styling classes to the form inputs and wrapping the form in a container with a background color. Additionally, unused code related to the logo animation and styling has been removed. These changes improve the visual appearance of the login form and remove unnecessary code clutter.

* 🔧 fix: update ESLint configuration to use double quotes consistently

* 💄 style: update sign up form layout

- Added a container div with flex gap for better alignment of form inputs.
- Wrapped the "User name" label and inputs in a div to keep them together.
- Added a closing div tag to close the container div.

* 🎨 style: update styling and placeholders in sign up form

- Updated the title to "ChatHub"
- Updated the styling of the form container
- Updated the styling and placeholders of the input fields
- Added a button for signing up