✨ Add error handling for server responses in login and register forms (#31)
* ✨ Add error handling for server responses in login and register forms

Added error handling for server responses in the login and register forms. Display
root error message on server message or server error. Display specific error
messages on form fields where necessary. Improves user experience in case of
server errors.

* ✨ Add more emphasis on the title in LogInPage component

Increased font size and added margin bottom to the title in the LogInPage component
by changing the classnames in the JSX file. This gives more importance to the title
"ChatHub" and improves the overall visual appeal of the page.

* ✨ Add registration page route

Added a new route for the registration page in the createBrowserRouter function to enable users to access the RegisterPage component. The registration page can now be accessed via "/register".

- Added new route for the registration page
- Connected the RegisterPage component to the new route